requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-cloud-bigquery==3.25.0
google-auth==2.34.0
